ramips: mt7621: mikrotik: initram use lzma loader
Use OpenWrt's lzma loader for initramfs image
Due to mips kernel self extractor inability to relocate,
and RouterBoot on 760igs refusing to netboot an ELF with:
Entry point address:               0x80b91000
or greater: 0x80b81000 and below okay.

Note that this is only a bootloader ELF entry point limitation,
a 50MiB Linux + initramfs vmlinux will boot fine on 760igs.

With this lzma loader, netboot will occasionally freeze after
decompressing the kernel, but before starting vmlinux. It was considered
that an occasional netboot freeze was simpler to workaround than an
image that refuses to boot unless minimized.

Signed-off-by: John Thomson <git@johnthomson.fastmail.com.au>